Add files via upload
A brief description of numpy.random and np.random.rand (x) and np.random.rand(a, b). Examples of same are given in the file along with a demonstration that np.random (x) does not generate any data and must be used along with another function. Examples of np.random.normal(x) however I'm still not sure what this does as it np.random.normal(50) which resulted in 54.87, np.random.normal(5) resulted in 6.105, They seem to produce a number close in proximity to (X) but are either slightly smaller or bigger than X.