Add files via upload
A description of 'arr = np.arange(x).reshape((a, b))   np.random.permutation(arr)' function and how the values of a and b effect the overall format of the result generated. 'a' represents the number of rows or while 'b' represents how many digits are in each each. 'a' and 'b' must evenly divide into 'x' in order for the function to run.